requirements.txt
elasticsearch==5.4.0
MySQL-python==1.2.5
mysql-replication==0.13
PyMySQL==0.7.11
PyYAML==3.12
requests==2.18.4
schedule==0.4.3
